#2 - dto 클래스 명 변경
기존에는  오픈 API 요청에 대한 응답을 'MedicineResponse'로 담았다.
하지만 이름이 명확하지 않고, 제품명으로 약을 조회 API를 개발할 때
'MedicineReponse'의 이름으로 응답 데이터를 담기에 더 적절하므로 클래스명을 변경

MedicineResponse => 약 조회 API에 사용하는 응답데이터 DTO
OpenApiMedicineResponse => 오픈 API의 응답데이터를 담기 위한 DTO
